Remove *Cng* APIs from netstandard 2.0
These are windows specific APIs and can be implemented on top of
netstandard similar to Registry so removing them from netstandard.
