--- netstandard/pkg/shims/netstandard/System.Security.Cryptography.Cng.csproj
@@ -1,14 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<Project ToolsVersion="12.0" DefaultTargets="Build" xmlns="http://schemas.microsoft.com/developer/msbuild/2003">
-  <Import Project="$([MSBuild]::GetDirectoryNameOfFileAbove($(MSBuildThisFileDirectory), dir.props))\dir.props" />
-  <PropertyGroup>
-    <AssemblyName>System.Security.Cryptography.Cng</AssemblyName>
-    <AssemblyVersion>4.2.0.0</AssemblyVersion>
-    <UseMSFTKey>true</UseMSFTKey>
-  </PropertyGroup>
-  <ItemGroup>
-    <Compile Include="System.Security.Cryptography.Cng.Forwards.cs" />
-    <ProjectReference Include="..\..\..\ref\netstandard.csproj" />
-  </ItemGroup>
-  <Import Project="$([MSBuild]::GetDirectoryNameOfFileAbove($(MSBuildThisFileDirectory), dir.targets))\dir.targets" />
-</Project>
